Put account balances into accounts, safer aeson instances
- Required a restructure because of cyclic imports.
- Handwrite aeson instances for important bits. Could do with tests
here.
- Put balances into accounts so we can use them as a proxy for creation
status.
