+++ common/src/Common/Foundation.hs
@@ -5,6 +5,7 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE PackageImports #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeFamilies #-}
 
@@ -15,6 +16,10 @@
 module Common.Foundation
   ( -- * Aeson encodings
     compactEncoding
+    -- * Lenses and Prisms
+  , makePactLenses
+  , makePactLensesNonClassy
+  , makePactPrisms
     -- * Helpers that should really not be here
   , tshow
   , prettyTextCompact
@@ -31,6 +36,7 @@ module Common.Foundation
   , module GHC.Generics
   ) where
 
+import           Control.Lens
 import           Control.Monad.Except                  (MonadError, throwError)
 import           Control.Monad.Fix
 import           Control.Monad.IO.Class
@@ -45,6 +51,7 @@ import qualified Data.Text                             as T
 import qualified Data.Text.Encoding                    as T
 import qualified Data.Text.Encoding.Error              as T
 import           GHC.Generics                          (Generic)
+import "template-haskell" Language.Haskell.TH
 
 import           Data.Maybe                            hiding (mapMaybe)
 
@@ -93,3 +100,41 @@ note e = maybe (throwError e) pure
 
 safeDecodeUtf8 :: ByteString -> Text
 safeDecodeUtf8 = T.decodeUtf8With T.lenientDecode
+
+-- | Lenses in this project should be generated by means of this function.
+--
+--   We generate lazy classy lenses. Classes make the export lists less tedious
+--   and allows for generic code, which will come in handy when the project
+--   grows.
+--
+--   We want lazy lenses so we can uses lenses also in recursive definitions.
+
+makePactLenses :: Name -> DecsQ
+makePactLenses =
+  makeLensesWith
+    ( classyRules
+        & generateLazyPatterns .~ True
+        & createClass .~ True
+    )
+
+-- | Non classy non simple lenses.
+--
+--   For some reason ( I have not investigated yet ), classy non simple lenses
+--   don't seem to work properly, at least not if the type has parameters.
+--   Therefore if you have a type where you need non simple lenses (lenses that
+--   can change the type), you have to use this function instead of
+--   `makePactLenses`.
+makePactLensesNonClassy :: Name -> DecsQ
+makePactLensesNonClassy =
+  makeLensesWith
+    ( lensRules
+        & simpleLenses .~ False
+        & generateLazyPatterns .~ True
+    )
+
+-- | Make Prisms in "pact style".
+--
+--   Currently this is just standard `makePrisms`
+makePactPrisms :: Name -> DecsQ
+makePactPrisms = makePrisms
+
