+++ common/src/Common/Network.hs
@@ -23,9 +23,6 @@ module Common.Network
   , getPactInstancePort
   , parseNetworks
   , Pact.ChainId (..)
-  , wrapWithBalanceChecks
-  , parseWrappedBalanceChecks
-  , AccountBalance(..)
   ) where
 
 import Control.Applicative ((<|>))
@@ -36,32 +33,18 @@ import Control.Monad
 import Control.Monad.Except (MonadError, liftEither, runExceptT, throwError, withExceptT)
 import Control.Monad.Trans
 import Data.Aeson
-import Data.Bifunctor (first)
 import Data.Coerce (coerce)
-import Data.Decimal (Decimal)
-import Data.Default
 import Data.Either (rights)
 import Data.Map (Map)
 import Data.Maybe (fromMaybe)
-import Data.Set (Set)
 import Data.Text (Text)
-import Data.Traversable (for)
 import Data.Void (Void)
 import GHC.Generics (Generic)
-import Kadena.SigningApi (AccountName(..), mkAccountName)
 import Obelisk.Configs
-import Pact.Compile (compileExps, mkEmptyInfo)
-import Pact.Parse
-import Pact.Types.Exp
-import Pact.Types.PactValue
-import Pact.Types.Pretty
-import Pact.Types.Term hiding (PublicKey)
-import Pact.Types.Type
 import Text.URI.Lens
 
 import qualified Data.Aeson as A
 import qualified Data.Map as Map
-import qualified Data.Set as Set
 import qualified Data.Text as T
 import qualified Pact.Types.ChainId as Pact
 import qualified Text.Megaparsec as MP
@@ -71,9 +54,6 @@ import Common.Api
 import Common.Orphans ()
 import Common.Foundation
 import Common.RefPath as MP
-
--- | Account balance wrapper
-newtype AccountBalance = AccountBalance { unAccountBalance :: Decimal } deriving (Eq, Ord, Num)
 
 -- | Name that uniquely describes a valid network.
 newtype NetworkName = NetworkName
@@ -230,82 +210,3 @@ renderNodeRef (NodeRef (URI.Authority mUser h mp)) =
     maybe "" ((<> "@") . renderUser) mUser <> URI.unRText h <> maybe "" ((":" <>) . tshow) mp
   where
     renderUser (URI.UserInfo name mPw) = URI.unRText name <> maybe "" ((":" <>) . URI.unRText) mPw
-
--- | Helper function for compiling pact code to a list of terms
-compileCode :: Text -> Either String [Term Name]
-compileCode = first show . compileExps mkEmptyInfo <=< parseExprs
-
--- | Parse the balance checking object into a map of account balance changes and
--- the result from the inner code
-parseWrappedBalanceChecks :: PactValue -> Either Text (Map AccountName AccountBalance, PactValue)
-parseWrappedBalanceChecks = first ("parseWrappedBalanceChecks: " <>) . \case
-  (PObject (ObjectMap obj)) -> do
-    let lookupErr k = case Map.lookup (FieldKey k) obj of
-          Nothing -> Left $ "Missing key '" <> k <> "' in map: " <> renderCompactText (ObjectMap obj)
-          Just v -> pure v
-    before <- parseAccountBalances =<< lookupErr "before"
-    result <- parseResults =<< lookupErr "results"
-    after <- parseAccountBalances =<< lookupErr "after"
-    pure (Map.unionWith subtract before after, result)
-  v -> Left $ "Unexpected PactValue (expected object): " <> renderCompactText v
-
--- | Turn the object of account->balance into a map
-parseAccountBalances :: PactValue -> Either Text (Map AccountName AccountBalance)
-parseAccountBalances = first ("parseAccountBalances: " <>) . \case
-  (PObject (ObjectMap obj)) -> do
-    m <- for (Map.toAscList obj) $ \(FieldKey accountText, pv) -> do
-      bal <- case pv of
-        PLiteral (LDecimal d) -> pure d
-        t -> Left $ "Unexpected PactValue (expected decimal): " <> renderCompactText t
-      acc <- mkAccountName accountText
-      pure (acc, AccountBalance bal)
-    pure $ Map.fromList m
-  v -> Left $ "Unexpected PactValue (expected object): " <> renderCompactText v
-
--- | Get the last result, as we would under unwrapped deployment.
-parseResults :: PactValue -> Either Text PactValue
-parseResults = first ("parseResults: " <>) . \case
-  PList vec -> maybe (Left "No value returned") Right $ vec ^? _last
-  v -> Left $ "Unexpected PactValue (expected list): " <> renderCompactText v
-
--- | Wrap the code with a let binding to get the balances of the given accounts
--- before and after executing the code.
-wrapWithBalanceChecks :: Set AccountName -> Text -> Either String Text
-wrapWithBalanceChecks accounts code = wrapped <$ compileCode code
-  where
-    getBalance :: AccountName -> (FieldKey, Term Name)
-    getBalance acc = (FieldKey (unAccountName acc), TApp
-      { _tApp = App
-        { _appFun = TVar (QName $ QualifiedName "coin" "get-balance" def) def
-        , _appArgs = [TLiteral (LString $ unAccountName acc) def]
-        , _appInfo = def
-        }
-      , _tInfo = def
-      })
-    -- Produce an object from account names to account balances
-    accountBalances = TObject
-      { _tObject = Pact.Types.Term.Object
-        { _oObject = ObjectMap $ Map.fromList $ map getBalance $ Set.toAscList accounts
-        , _oObjectType = TyPrim TyDecimal
-        , _oKeyOrder = Nothing
-        , _oInfo = def
-        }
-      , _tInfo = def
-      }
-    -- It would be nice to parse and compile the code and shove it into a
-    -- giant 'Term' so we can serialise it, but 'pretty' is not guaranteed to
-    -- produce valid pact code. It at least produces bad type sigs and let
-    -- bindings.
-    -- Thus we build this from a string and use 'pretty' where appropriate.
-    -- We do need to make sure the code compiles before splicing it in.
-    -- The order of execution is the same as the order of the bound variables.
-    wrapped = T.unlines
-      [ "(let"
-      , "  ((before " <> renderCompactText accountBalances <> ")"
-      , "   (results ["
-      , code
-      , "   ])"
-      , "   (after " <> renderCompactText accountBalances <> "))"
-      , "   {\"after\": after, \"results\": results, \"before\": before})"
-      ]
-
