+++ frontend/src/Frontend/Foundation.hs
@@ -19,10 +19,6 @@ module Frontend.Foundation
     ReflexValue
   , MDynamic
   , LeftmostEv (..)
-    -- * Lenses and Prisms
-  , makePactLenses
-  , makePactLensesNonClassy
-  , makePactPrisms
     -- * Helpers that should really not be here
   , getBrowserProperty
     -- * Common Foundation
@@ -44,8 +40,6 @@ import           Data.Foldable
 import           Data.Semigroup
 import           Data.Text                         (Text)
 import           GHC.Generics                      (Generic)
-import "template-haskell" Language.Haskell.TH               (DecsQ)
-import "template-haskell" Language.Haskell.TH.Syntax        (Name)
 import           Language.Javascript.JSaddle       (JSM, MonadJSM, askJSM,
                                                     liftJSM, runJSM)
 import qualified Language.Javascript.JSaddle       as JS
@@ -80,43 +74,6 @@ instance Reflex t => Monoid (LeftmostEv t a) where
   mempty = LeftmostEv never
   mappend = (<>)
   mconcat = LeftmostEv . leftmost . coerce
-
--- | Lenses in this project should be generated by means of this function.
---
---   We generate lazy classy lenses. Classes make the export lists less tedious
---   and allows for generic code, which will come in handy when the project
---   grows.
---
---   We want lazy lenses so we can uses lenses also in recursive definitions.
-
-makePactLenses :: Name -> DecsQ
-makePactLenses =
-  makeLensesWith
-    ( classyRules
-        & generateLazyPatterns .~ True
-        & createClass .~ True
-    )
-
--- | Non classy non simple lenses.
---
---   For some reason ( I have not investigated yet ), classy non simple lenses
---   don't seem to work properly, at least not if the type has parameters.
---   Therefore if you have a type where you need non simple lenses (lenses that
---   can change the type), you have to use this function instead of
---   `makePactLenses`.
-makePactLensesNonClassy :: Name -> DecsQ
-makePactLensesNonClassy =
-  makeLensesWith
-    ( lensRules
-        & simpleLenses .~ False
-        & generateLazyPatterns .~ True
-    )
-
--- | Make Prisms in "pact style".
---
---   Currently this is just standard `makePrisms`
-makePactPrisms :: Name -> DecsQ
-makePactPrisms = makePrisms
 
 getBrowserProperty :: forall m. MonadJSM m => Text -> m Bool
 getBrowserProperty property = liftJSM $ fromMaybe False <$> JS.catch (JS.fromJSVal =<< JS.eval ("bowser." <> property)) (\(_ :: JS.JSException) -> pure Nothing)
