+++ frontend/src/Frontend/UI/Wallet.hs
@@ -18,10 +18,6 @@ module Frontend.UI.Wallet
   ( -- * Key management widget
     uiWallet
   , uiAvailableKeys
-    -- * Keys related helper widgets
---  , uiSelectKey
-  , getBalance
-  , showBalance
     -- ** Filters for keys
   , hasPrivateKey
   , HasUiWalletModelCfg
@@ -33,12 +29,14 @@ import           Control.Monad               (when, (<=<))
 import qualified Data.IntMap                 as IntMap
 import qualified Data.Map                    as Map
 import           Data.Text                   (Text)
+import Obelisk.Route.Frontend
 import qualified Pact.Types.PactValue as Pact
 import qualified Pact.Types.Exp as Pact
 import qualified Pact.Types.ChainId as Pact
 import           Reflex
 import           Reflex.Dom
 ------------------------------------------------------------------------------
+import Common.Route
 import           Frontend.Crypto.Class
 import           Frontend.Crypto.Ed25519     (keyToText)
 import           Frontend.Wallet
@@ -82,6 +80,7 @@ uiWallet
      . ( MonadWidget t m
        , HasUiWalletModelCfg model mConf key m t
        , HasCrypto key m
+       , Routed t (R FrontendRoute) m
        )
   => model
   -> m mConf
@@ -103,6 +102,7 @@ uiAvailableKeys
      ( MonadWidget t m
      , HasUiWalletModelCfg model mConf key m t
      , HasCrypto key m
+     , Routed t (R FrontendRoute) m
      )
   => model
   -> m mConf
@@ -119,6 +119,7 @@ uiKeyItems
      ( MonadWidget t m
      , HasUiWalletModelCfg model mConf key m t
      , HasCrypto key m
+     , Routed t (R FrontendRoute) m
      )
   => model
   -> m mConf
@@ -161,7 +162,16 @@ uiKeyItems model = do
       AccountDialog_Receive -> uiReceiveModal model a
       AccountDialog_Send -> uiSendModal model a
 
-  pure $ mempty & modalCfg_setModal .~ (accModal <$> onAccountModal)
+  dr <- askRoute
+  let refresh = fforMaybe (updated dr) $ \case
+        FrontendRoute_Wallet :/ () -> Just ()
+        _ -> Nothing
+
+  rbtn <- button "refresh"
+
+  pure $ mempty
+    & modalCfg_setModal .~ (accModal <$> onAccountModal)
+    & walletCfg_refreshBalances .~ leftmost [refresh, rbtn]
 
 ------------------------------------------------------------------------------
 -- | Display a key as list item together with it's name.
@@ -188,7 +198,9 @@ uiKeyItem model i d = do
             td $ divClass "wallet__table-wallet-address" $ dynText $ keyToText . _keyPair_publicKey . _account_key <$> account
             td $ dynText $ Pact._chainId . _account_chainId <$> account
             td $ dynText $ _account_notes <$> account
-            td $ dyn_ $ ffor2 (_account_chainId <$> account) (_account_name <$> account) (showBalance model never)
+            td $ dynText $ ffor account $ \a -> case _account_balance a of
+              Nothing -> "Unknown"
+              Just b -> tshow (unAccountBalance b) <> " KDA"
 
             td $ divClass "wallet__table-buttons" $ do
               let cfg = def
@@ -205,33 +217,3 @@ uiKeyItem model i d = do
                 , mkDialog AccountDialog_Receive recv
                 , mkDialog AccountDialog_Send send
                 ]
-
--- | Get the balance of an account from the network. 'Nothing' indicates _some_
--- failure, either a missing account or connectivity failure. We have no need to
--- distinguish between the two at this point.
-getBalance
-  :: (MonadWidget t m, HasNetwork model t, HasCrypto key (Performable m))
-  => model -> ChainId -> Event t AccountName -> m (Event t (Maybe AccountBalance))
-getBalance model chain account = do
-  networkRequest <- performEvent $ attachWith mkReq (current $ getNetworkNameAndMeta model) account
-  response <- performLocalReadCustom (model ^. network) pure networkRequest
-  let toBalance (_, [Right (_, Pact.PLiteral (Pact.LDecimal d))]) = Just $ AccountBalance d
-      toBalance _ = Nothing
-  pure $ toBalance <$> response
-  where
-    accountBalanceReq acc = "(coin.get-balance " <> tshow (unAccountName acc) <> ")"
-    mkReq (netName, pm) acc = mkSimpleReadReq (accountBalanceReq acc) netName pm (ChainRef Nothing chain)
-
--- | Display the balance of an account after retrieving it from the network
-showBalance
-  :: (MonadWidget t m, HasNetwork model t, HasCrypto key (Performable m))
-  => model -> Event t () -> ChainId -> AccountName -> m ()
-showBalance model refresh chain acc = do
-  -- This delay ensures we have the networks stuff set up by the time we do the
-  -- requests, thus avoiding immediate failure.
-  pb <- delay 2 =<< getPostBuild
-  bal <- getBalance model chain $ acc <$ (pb <> refresh)
-  _ <- runWithReplace (text "Loading...") $ ffor bal $ \case
-    Nothing -> text "Unknown"
-    Just b -> text $ tshow (unAccountBalance b) <> " KDA"
-  pure ()
